WIP update setConsent to use extensionViews and transientViews
