+++ src/view/actions/setConsent.jsx
@@ -12,228 +12,41 @@ governing permissions and limitations under the License.
 
 import "regenerator-runtime"; // needed for some of react-spectrum
 import React, { Fragment } from "react";
-import { object, string, array, mixed } from "yup";
+import { object, string } from "yup";
 
-import { Form, Item, Radio, Well, Button } from "@adobe/react-spectrum";
+import { Form, Item, Button } from "@adobe/react-spectrum";
 import Delete from "@spectrum-icons/workflow/Delete";
 import {
   TextField,
-  RadioGroup,
   Picker
 } from "../components/formikReactSpectrum3";
-import OptionsWithDataElement from "../components/formikReactSpectrum3/optionsWithDataElement";
+//import OptionsWithDataElement from "../components/formikReactSpectrum3/optionsWithDataElement";
 import DataElementSelector from "../components/dataElementSelector";
 import render from "../spectrum3Render";
-
-import PropTypes from "prop-types";
-//import { FieldArray } from "formik";
 
 import ExtensionView from "../components/spectrum3ExtensionView";
 import getInstanceOptions from "../utils/getInstanceOptions";
 import singleDataElementRegex from "../constants/singleDataElementRegex";
-//import "./setConsent.styl";
-import {
-  DATA_ELEMENT as DATA_ELEMENT_OPTION
-} from "../components/optionsWithDataElement";
-import { useField, FieldArray } from "formik";
-
-const IN = { value: "in", label: "In" };
-const OUT = { value: "out", label: "Out" };
-const FORM = { value: "form", label: "Fill out a form" };
-const DATA_ELEMENT = { value: "dataElement", label: "Use a data element" };
-const YES = { value: "yes", label: "Yes" };
-const NO = { value: "no", label: "No" };
-const ADOBE = { value: "adobe", label: "Adobe" };
-const IAB_TCF = { value: "iab_tcf", label: "IAB TCF" };
-const VERSION_1_0 = { value: "1.0", label: "1.0" };
-const VERSION_2_0 = { value: "2.0", label: "2.0" };
-
-/*
- * The settings object for this action has the form of:
- * {
- *   instanceName, // the name of the instance
- *   consent: [{
- *    standard: "Adobe",
- *    version: "1.0",
- *    value: {
- *      general // valid values are "in" and "out"
- *    }
- *   },
- *   {
- *     standard: "IAB TCF",
- *     version: "2.0",
- *     value: "1234abcd",
- *     gdprApplies: true,
- *     gdprContainsPersonalData: false
- *   },
- *   {
- *     standard: "Adobe",
- *     version: "2.0",
- *     value: "%dataElement123%"
- *   }
- *   ...]
- * }
- * ---OR---
- * {
- *   instanceName,
- *   // this should resolve to the preferences object above
- *   consent: "%dataElement123%"
- * }
- *
- * The formik object looks like this:
- * {
- *   instanceName,
- *   inputMethod, // "dataElement" or "form"
- *   dataElement: "%data_element%" or
- *   consent: [
- *     {
- *       standard: ("adobe" or "iab_tcf"),
- *       adobeVersion: "1.0",
- *       iabVersion: "2.0",
- *       general: optionsWithDataElement("in", "out"),
- *       value: "%data_element%"
- *       iabValue: "1234abcd", // or data_element
- *       gdprApplies: optionsWithDataElement(true, false),
- *       gdprContainsPersonalData: optionsWithDataElement(true, false)
- *     },
- *     ...
- *   ]
- * }
- *
- * When we get to more granular purposes, we can augment the settings object to include
- * additional purposes. We will need to re-vamp the UI and the formik object when that happens.
- */
-
-const createBlankConsentObject = () => {
-  return {
-    standard: "adobe",
-    adobeVersion: "1.0",
-    iabVersion: "2.0",
-    general: { radio: IN.value, dataElement: "" },
-    value: "",
-    iabValue: "",
-    gdprApplies: { radio: YES.value, dataElement: "" },
-    gdprContainsPersonalData: { radio: NO.value, dataElement: "" }
-  };
-};
-
-const applyInitialValuesForOptionsWithDataElement = (
-  value,
-  formikValues,
-  options
-) => {
-  if (singleDataElementRegex.test(value)) {
-    formikValues.radio = DATA_ELEMENT_OPTION;
-    formikValues.dataElement = value;
-  } else if (options.includes(value)) {
-    formikValues.radio = value;
-  }
-};
-
-const booleanToYesNo = value => {
-  if (value === true) {
-    return YES.value;
-  }
-  if (value === false) {
-    return NO.value;
-  }
-  // in this case it most likely is a dataElement string
-  return value;
-};
+import ExtensionViewForm from "../components/extensionViewForm";
+import DataElementRadioChoice from "../components/dataElementRadioChoice";
+import ConsentObjects from "../components/consentObjects";
 
 const getInitialValues = ({ initInfo }) => {
   const {
     instanceName = initInfo.extensionSettings.instances[0].name,
-    consent,
-    identityMap
+    identityMap = ""
   } = initInfo.settings || {};
 
-  const initialValues = {
+  return {
     instanceName,
     identityMap
   };
-
-  if (typeof consent === "string") {
-    initialValues.inputMethod = DATA_ELEMENT.value;
-    initialValues.dataElement = consent;
-    initialValues.consentObjects = [createBlankConsentObject()];
-  } else if (Array.isArray(consent)) {
-    initialValues.inputMethod = FORM.value;
-    initialValues.consent = consent.reduce((memo, consentObject) => {
-      const formikConsentObject = createBlankConsentObject();
-      if (consentObject.standard === ADOBE.label) {
-        formikConsentObject.standard = ADOBE.value;
-        if (consentObject.version) {
-          formikConsentObject.adobeVersion = consentObject.version;
-        }
-        if (consentObject.value && consentObject.value.general) {
-          applyInitialValuesForOptionsWithDataElement(
-            consentObject.value.general,
-            formikConsentObject.general,
-            [IN.value, OUT.value]
-          );
-        } else if (consentObject.value) {
-          formikConsentObject.value = consentObject.value;
-        }
-      } else if (consentObject.standard === IAB_TCF.label) {
-        formikConsentObject.standard = IAB_TCF.value;
-        if (consentObject.version) {
-          formikConsentObject.iabVersion = consentObject.version;
-        }
-        if (consentObject.value) {
-          formikConsentObject.iabValue = consentObject.value;
-        }
-        if (consentObject.gdprApplies != null) {
-          applyInitialValuesForOptionsWithDataElement(
-            booleanToYesNo(consentObject.gdprApplies),
-            formikConsentObject.gdprApplies,
-            [YES.value, NO.value]
-          );
-        }
-        if (consentObject.gdprContainsPersonalData != null) {
-          applyInitialValuesForOptionsWithDataElement(
-            booleanToYesNo(consentObject.gdprContainsPersonalData),
-            formikConsentObject.gdprContainsPersonalData,
-            [YES.value, NO.value]
-          );
-        }
-      }
-      memo.push(formikConsentObject);
-      return memo;
-    }, []);
-  } else {
-    initialValues.inputMethod = FORM.value;
-    initialValues.dataElement = "";
-    initialValues.consent = [createBlankConsentObject()];
-  }
-  return initialValues;
-};
-
-const getSettingsForOptionsWithDataElement = value => {
-  if (value.radio === DATA_ELEMENT_OPTION) {
-    return value.dataElement;
-  }
-  return value.radio;
-};
-
-const yesNoToBoolean = value => {
-  if (value === YES.value) {
-    return true;
-  }
-  if (value === NO.value) {
-    return false;
-  }
-  // handle dataElement strings
-  return value;
 };
 
 const getSettings = ({ values }) => {
   const {
     instanceName,
-    identityMap,
-    inputMethod,
-    dataElement,
-    consent
+    identityMap
   } = values;
 
   const settings = {
@@ -244,460 +57,62 @@ const getSettings = ({ values }) => {
     settings.identityMap = identityMap;
   }
 
-  if (inputMethod === DATA_ELEMENT.value) {
-    settings.consent = dataElement;
-  } else {
-    settings.consent = consent.reduce((memo, formikConsentObject) => {
-      if (formikConsentObject.standard === ADOBE.value) {
-        if (formikConsentObject.adobeVersion === "1.0") {
-          memo.push({
-            standard: ADOBE.label,
-            version: formikConsentObject.adobeVersion,
-            value: {
-              general: getSettingsForOptionsWithDataElement(
-                formikConsentObject.general
-              )
-            }
-          });
-        } else {
-          memo.push({
-            standard: ADOBE.label,
-            version: formikConsentObject.adobeVersion,
-            value: formikConsentObject.value
-          });
-        }
-      } else if (formikConsentObject.standard === IAB_TCF.value) {
-        memo.push({
-          standard: IAB_TCF.label,
-          version: formikConsentObject.iabVersion,
-          value: formikConsentObject.iabValue,
-          gdprApplies: yesNoToBoolean(
-            getSettingsForOptionsWithDataElement(
-              formikConsentObject.gdprApplies
-            )
-          ),
-          gdprContainsPersonalData: yesNoToBoolean(
-            getSettingsForOptionsWithDataElement(
-              formikConsentObject.gdprContainsPersonalData
-            )
-          )
-        });
-      }
-      return memo;
-    }, []);
-  }
-
   return settings;
-};
+}
 
 const invalidDataMessage = "Please specify a data element.";
 const validationSchema = object().shape({
   instanceName: string().required(),
-  identityMap: string().matches(singleDataElementRegex, invalidDataMessage),
-  dataElement: mixed().when("inputMethod", {
-    is: DATA_ELEMENT.value,
-    then: string()
-      .matches(singleDataElementRegex, invalidDataMessage)
-      .required(invalidDataMessage)
-  }),
-  consent: array().when("inputMethod", {
-    is: FORM.value,
-    then: array().of(
-      object().shape({
-        standard: string().required("Plesase specify a standard."),
-        general: mixed().when(["standard", "adobeVersion"], {
-          is: (standard, adobeVersion) =>
-            standard === ADOBE.value && adobeVersion === "1.0",
-          then: object().shape({
-            dataElement: mixed().when("radio", {
-              is: DATA_ELEMENT_OPTION,
-              then: string().matches(singleDataElementRegex, invalidDataMessage)
-            })
-          })
-        }),
-        value: mixed().when(["standard", "adobeVersion"], {
-          is: (standard, adobeVersion) =>
-            standard === ADOBE.value && adobeVersion !== "1.0",
-          then: string()
-            .required()
-            .matches(singleDataElementRegex, invalidDataMessage)
-        }),
-        iabVersion: mixed().when(["standard"], {
-          is: IAB_TCF.value,
-          then: string().required("Please specify a version.")
-        }),
-        iabValue: mixed().when("standard", {
-          is: IAB_TCF.value,
-          then: string().required("Please specify a value.")
-        }),
-        gdprApplies: mixed().when("standard", {
-          is: IAB_TCF.value,
-          then: object().shape({
-            radio: string().required(),
-            dataElement: mixed().when("radio", {
-              is: DATA_ELEMENT_OPTION,
-              then: string().matches(singleDataElementRegex, invalidDataMessage)
-            })
-          })
-        }),
-        gdprContainsPersonalData: mixed().when("standard", {
-          is: IAB_TCF.value,
-          then: object().shape({
-            radio: string().required(),
-            dataElement: mixed().when("radio", {
-              is: DATA_ELEMENT_OPTION,
-              then: string().matches(singleDataElementRegex, invalidDataMessage)
-            })
-          })
-        })
-      })
-    )
-  })
+  identityMap: string().matches(singleDataElementRegex, invalidDataMessage)
 });
-
-const ConsentObject = ({ formikConsentObject, index }) => {
-  return (
-    <Fragment>
-      <Picker
-        data-test-id="standardSelect"
-        name={`consent.${index}.standard`}
-        label="Consent Standard"
-        items={[ADOBE, IAB_TCF]}
-        width="size-5000"
-        isRequired
-      >
-        {item => <Item key={item.value}>{item.label}</Item>}
-      </Picker>
-      {formikConsentObject.standard === ADOBE.value && (
-        <Picker
-          data-test-id="adobeVersionSelect"
-          name={`consent.${index}.adobeVersion`}
-          label="Consent Version"
-          items={[VERSION_1_0, VERSION_2_0]}
-          width="size-5000"
-          isRequired
-        >
-          {item => <Item key={item.value}>{item.label}</Item>}
-        </Picker>
-      )}
-      <OptionsWithDataElement
-        name="foo"
-        width="size-5000"
-        />
-    </Fragment>
-  )
-};
-/*
-
-      {formikConsentObject.standard === ADOBE.value &&
-        formikConsentObject.adobeVersion === "1.0" && (
-          <OptionsWithDataElement
-            label="General Consent"
-            infoTip="The general consent level. If provided through a data element, it should resolve to 'in' or 'out'."
-            id={`consent_${index}_general`}
-            data-test-id="general"
-            name={`consent.${index}.general`}
-            options={[IN, OUT]}
-            values={formikConsentObject.general}
-          />
-        )}
-      {formikConsentObject.standard === ADOBE.value &&
-        formikConsentObject.adobeVersion !== "1.0" && (
-          <Fragment>
-            <InfoTipLayout tip="A data element containing the Adobe consent XDM object">
-              <FieldLabel labelFor={`consent_${index}_value`} label="Value" />
-            </InfoTipLayout>
-            <WrappedField
-              data-test-id="valueField"
-              id={`consent_${index}_value`}
-              name={`consent.${index}.value`}
-              component={Textfield}
-              componentClassName="u-fieldLong"
-              supportDataElement="replace"
-            />
-          </Fragment>
-        )}
-      {formikConsentObject.standard === IAB_TCF.value && (
-        <div>
-          <div className="u-gapTop">
-            <InfoTipLayout tip="The consent standard version for this consent object">
-              <FieldLabel
-                labelFor={`consent.${index}.iabVersion`}
-                label="Version"
-              />
-            </InfoTipLayout>
-            <WrappedField
-              data-test-id="iabVersionField"
-              id={`consent.${index}.iabVersion`}
-              name={`consent.${index}.iabVersion`}
-              component={Textfield}
-              componentClassName="u-fieldLong"
-            />
-          </div>
-          <div className="u-gapTop">
-            <InfoTipLayout tip="The encoded IAB TCF consent value">
-              <FieldLabel
-                labelFor={`consent_${index}_iabValue`}
-                label="Value"
-              />
-            </InfoTipLayout>
-            <WrappedField
-              data-test-id="iabValueField"
-              id={`consent_${index}_iabValue`}
-              name={`consent.${index}.iabValue`}
-              component={Textfield}
-              componentClassName="u-fieldLong"
-              supportDataElement="append"
-            />
-          </div>
-          <OptionsWithDataElement
-            label="GDPR Applies"
-            infoTip="Does GDPR apply to this consent value? A data element should resolve to true or false."
-            id={`consent_${index}_gdprApplies`}
-            data-test-id="gdprApplies"
-            name={`consent.${index}.gdprApplies`}
-            options={[YES, NO]}
-            values={formikConsentObject.gdprApplies}
-          />
-          <OptionsWithDataElement
-            label="GDPR Contains Personal Data"
-            infoTip="Does the event data associated with this user contain personal data? A data element should resolve to true or false."
-            id={`consent_${index}_gdprContainsPersonalData`}
-            data-test-id="gdprContainsPersonalData"
-            name={`consent.${index}.gdprContainsPersonalData`}
-            options={[YES, NO]}
-            values={formikConsentObject.gdprContainsPersonalData}
-          />
-        </div>
-      )}
-    </div>
-  );
-};
-*/
-ConsentObject.propTypes = {
-  formikConsentObject: PropTypes.object,
-  index: PropTypes.number
-};
 
 const SetConsent = () => {
 
   return (
     <ExtensionView
-      getInitialValues={getInitialValues}
-      getSettings={getSettings}
-      validationSchema={validationSchema}
-      render={({ initInfo, formikProps }) => {
-        const { values } = formikProps;
-        return (
-          <Form>
-            <Picker
-              data-test-id="instanceNameField"
-              name="instanceName"
-              label="Instance"
-              items={getInstanceOptions(initInfo)}
-              width="size-5000"
-              isRequired
-            >
-              {item => <Item key={item.value}>{item.label}</Item>}
-            </Picker>
-            <DataElementSelector>
-              <TextField
-                data-test-id="identityMapField"
-                name="identityMap"
-                label="Identity Map"
-                description="Provide a data element which returns a custom identity map object as part of the setConsent command."
-                width="size-5000"
-              />
-            </DataElementSelector>
-            <RadioGroup
-                name="inputMethod"
-                orientation="horizontal"
-                label="Consent Information"
-            >
-              <Radio data-test-id="inputMethodFormRadio" value={FORM.value}>
-                {FORM.label}
-              </Radio>
-              <Radio data-test-id="inputMethodDataElementRadio" value={DATA_ELEMENT.value}>
-                {DATA_ELEMENT.label}
-              </Radio>
-            </RadioGroup>
-            {values.inputMethod === FORM.value && (
-              <FieldArray
-                name="consent"
-                render={arrayHelpers => (
-                  <Fragment>
-                    <div className="u-gapTop u-alignRight">
-                      <Button
-                        variant="secondary"
-                        data-test-id="addConsentButton"
-                        onPress={() => {
-                          arrayHelpers.push(createBlankConsentObject());
-                        }}
-                      >
-                        Add Consent Object
-                      </Button>
-                    </div>
-                    {values.consent.map((formikConsentObject, index) => (
-                      <Fragment key={index}>
-                        <ConsentObject
-                          formikConsentObject={formikConsentObject}
-                          index={index}
-                        />
-                        <Button
-                          data-test-id="deleteConsentButton"
-                          isQuiet
-                          isDisabled={values.consent.length === 1}
-                          variant="secondary"
-                          onPress={() => {
-                            arrayHelpers.remove(index);
-                          }}
-                          aria-label="Delete Consent Object"
-                          UNSAFE_className="u-verticalCenterBottom"
-                          minWidth={0}
-                        >
-                          <Delete />
-                        </Button>
-                      </Fragment>
-                    ))}
-                  </Fragment>
-                )}
-              />
-            )}
-            {values.inputMethod === DATA_ELEMENT.value && (
-              <DataElementSelector>
-                <TextField
-                  data-test-id="dataElementField"
-                  name="dataElement"
-                  width="size-5000"
-                />
-              </DataElementSelector>
-            )}
-          </Form>
-        );
-      }}
-    />
-  );
-};
-/*          <div>
-            <div>
-              <FieldLabel labelFor="instanceNameField" label="Instance" />
-              <div>
-                <WrappedField
-                  data-test-id="instanceNameSelect"
-                  id="instanceNameField"
+      render={({ initInfo }) => (
+        <ExtensionViewForm
+          initialValues={getInitialValues({ initInfo })}
+          getSettings={getSettings}
+          validationSchema={validationSchema}
+          render={() => {
+            return (
+              <Form>
+                <Picker
+                  data-test-id="instanceNameField"
                   name="instanceName"
-                  component={Select}
-                  componentClassName="u-fieldLong"
-                  options={getInstanceOptions(initInfo)}
-                />
-              </div>
-            </div>
-            <div className="u-gapTop">
-              <InfoTipLayout tip="Use a data element to provide custom identity information as part of the setConsent command. This data element should resolve to an identity map object.">
-                <FieldLabel
-                  labelFor="identityMapField"
-                  label="Identity Map (Optional)"
-                />
-              </InfoTipLayout>
-              <WrappedField
-                data-test-id="identityMapField"
-                id="identityMapField"
-                name="identityMap"
-                component={Textfield}
-                componentClassName="u-fieldLong"
-                supportDataElement="replace"
-              />
-            </div>
-            <div className="u-gapTop">
-              <InfoTipLayout
-                tip="You can use a guided form to enter the consent information,
-                     or provide a data element."
-              >
-                <Heading variant="subtitle2">Consent Information</Heading>
-              </InfoTipLayout>
-              <WrappedField
-                name="inputMethod"
-                id="inputMethod"
-                component={RadioGroup}
-              >
-                <Radio
-                  data-test-id="inputMethodFormRadio"
-                  value={FORM.value}
-                  label={FORM.label}
-                />
-                <Radio
-                  data-test-id="inputMethodDataElementRadio"
-                  value={DATA_ELEMENT.value}
-                  label={DATA_ELEMENT.label}
-                  className="u-gapLeft2x"
-                />
-              </WrappedField>
-            </div>
-            {values.inputMethod === FORM.value && (
-              <div>
-                <FieldArray
+                  label="Instance"
+                  items={getInstanceOptions(initInfo)}
+                  width="size-5000"
+                  isRequired
+                >
+                  {item => <Item key={item.value}>{item.label}</Item>}
+                </Picker>
+                <DataElementSelector>
+                  <TextField
+                    data-test-id="identityMapField"
+                    name="identityMap"
+                    label="Identity Map"
+                    description="Provide a data element which returns a custom identity map object as part of the setConsent command."
+                    width="size-5000"
+                  />
+                </DataElementSelector>
+                <DataElementRadioChoice
                   name="consent"
-                  render={arrayHelpers => (
-                    <Fragment>
-                      <div className="u-gapTop u-alignRight">
-                        <Button
-                          data-test-id="addConsentButton"
-                          label="Add Consent Object"
-                          icon={<Add />}
-                          onClick={() => {
-                            arrayHelpers.push(createBlankConsentObject());
-                          }}
-                        />
-                      </div>
-                      {values.consent.map((formikConsentObject, index) => (
-                        <Well
-                          key={`consent${index}`}
-                          className="u-gapTop"
-                          data-test-id={`instance${index}`}
-                        >
-                          <ConsentObject
-                            formikConsentObject={formikConsentObject}
-                            index={index}
-                          />
-                          <Button
-                            data-test-id="deleteConsentButton"
-                            label="Delete Consent Object"
-                            icon={<Delete />}
-                            variant="action"
-                            disabled={values.consent.length === 1}
-                            onClick={() => {
-                              arrayHelpers.remove(index);
-                            }}
-                            className="u-gapTop"
-                          />
-                        </Well>
-                      ))}
-                    </Fragment>
-                  )}
-                />
-              </div>
-            )}
-            {values.inputMethod === DATA_ELEMENT.value && (
-              <div className="u-gapTop">
-                <InfoTipLayout tip="A data element containing an array of consent objects">
-                  <FieldLabel labelFor="dataElement" label="Data Element" />
-                </InfoTipLayout>
-                <WrappedField
-                  data-test-id="dataElementField"
-                  id="dataElement"
-                  name="dataElement"
-                  component={Textfield}
-                  componentClassName="u-fieldLong"
-                  supportDataElement="replace"
-                />
-              </div>
-            )}
-          </div>
-        );
-      }}
+                  label="Consent Information"
+                  dataElementLabel="Use a data element"
+                  constantLabel="Fill out a form"
+                  isRequired
+                >
+                  <ConsentObjects/>
+                </DataElementRadioChoice>
+              </Form>
+            );
+          }}
+        />
+      )}
     />
   );
 };
-*/
 
 render(SetConsent);
